Flutter Chat App 2
This app is a #ui messaging/chat app that runs both Android and iOS devices using #flutter and also has a dark and light theme. I created a total of 4 screens that support both Dark Theme and Light Theme. First, I designed a welcome screen with a slogan and there is a skip button. Once the user presses it, they are transferred to the Sign in or Sign up screen. While the last two screens are the main screen, the Chat screens show the list of people you are chatting with. End message screen supporting text, voice and video messages.